refactor: bare metal exoprt in `hooks` folder
